数据对齐：写入当日收藏变更

+++ xhs-data-align/src/main/java/org/n1vnhil/xhs/data/align/job/CreateTableXxlJob.java
@@ -28,7 +28,7 @@ public class CreateTableXxlJob {
         XxlJobHelper.log("## 开始初始化明日增量数据表");
 
         if(tableShards > 0) {
-            for(int i = 0; i < tableShards; i++) {
+            for(long i = 0; i < tableShards; i++) {
                 String tableNameSuffix = TableConstants.buildTableNameSuffix(date, i);
                 createTableMapper.createDataAlignFollowingCountTemplateTable(tableNameSuffix);
                 createTableMapper.createDataAlignNoteCollectCountTemplateTable(tableNameSuffix);
